# api/swagger.yaml
openapi: 3.0.0
info:
  title: Marvel API Client
  description: This is a simple Client for fetching Marvel Characters.
  version: 1.0.3-oas3
servers:
- url: /
tags:
- name: public
paths:
  /characters:
    get:
      tags:
      - public
      summary: Fetches lists of characters.
      description: Fetches lists of comic characters with optional filters. See notes on individual parameters below.
      parameters:
      - name: limit
        in: query
        description: Limit the result set to the specified number of resources.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        description: Skip the specified number of resources in the result set.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: No response was specified
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CharacterDataWrapper'
  /characters/{characterId}:
    get:
      tags:
      - public
      summary: Fetches a single character by id.
      description: This method fetches a single character resource.  It is the canonical URI for any character resource provided by the API.
      parameters:
      - name: characterId
        in: path
        description: A single character id.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: No response was specified
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CharacterDataWrapper'
  /comics:
    get:
      tags:
      - public
      summary: Fetches lists of comics.
      description: Fetches lists of comics with optional filters. See notes on individual parameters below.
      parameters:
      - name: limit
        in: query
        description: Limit the result set to the specified number of resources.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        description: Skip the specified number of resources in the result set.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: No response was specified
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComicDataWrapper'
  /comics/{comicId}:
    get:
      tags:
      - public
      summary: Fetches a single comic by id.
      description: This method fetches a single comic resource.  It is the canonical URI for any comic resource provided by the API.
      parameters:
      - name: comicId
        in: path
        description: A single comic.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: No response was specified
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ComicDataWrapper'
  /comics/{comicId}/characters:
    get:
      tags:
      - public
      summary: Fetches lists of characters filtered by a comic id.
      description: Fetches lists of characters which appear in a specific comic with optional filters. See notes on individual parameters below.
      parameters:
      - name: comicId
        in: path
        description: A single comic.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: Limit the result set to the specified number of resources.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        description: Skip the specified number of resources in the result set.
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: No response was specified
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CharacterDataWrapper'
components:
  schemas:
    ComicPrice:
      properties:
        type:
          type: string
          description: A description of the price (e.g. print price, digital price).
        price:
          type: number
          format: float
          description: The price (all prices in USD).
    ComicDate:
      properties:
        type:
          type: string
          description: A description of the date (e.g. onsale date, FOC date).
        date:
          type: string
          format: date
          description: The date.
    TextObject:
      properties:
        type:
          type: string
          description: The canonical type of the text object (e.g. solicit text, preview text, etc.).
        language:
          type: string
          description: The IETF language tag denoting the language the text object is written in.
        text:
          type: string
          description: The text.
    ComicList:
      properties:
        available:
          type: integer
          description: The number of total available issues in this list. Will always
            be greater than or equal to the "returned" value.
          format: int32
        returned:
          type: integer
          description: The number of issues returned in this collection (up to 20).
          format: int32
        collectionURI:
          type: string
          description: The path to the full list of issues in this collection.
        items:
          type: array
          description: The list of returned issues in this collection.
          items:
            $ref: '#/components/schemas/ComicSummary'
    EventList:
      properties:
        available:
          type: integer
          description: The number of total available events in this list. Will always
            be greater than or equal to the "returned" value.
          format: int32
        returned:
          type: integer
          description: The number of events returned in this collection (up to 20).
          format: int32
        collectionURI:
          type: string
          description: The path to the full list of events in this collection.
        items:
          type: array
          description: The list of returned events in this collection.
          items:
            $ref: '#/components/schemas/EventSummary'
    CreatorList:
      properties:
        available:
          type: integer
          description: The number of total available creators in this list. Will always be greater than or equal to the "returned" value.
          format: int32
        returned:
          type: integer
          description: The number of creators returned in this collection (up to 20).
          format: int32
        collectionURI:
          type: string
          description: The path to the full list of creators in this collection.
        items:
          type: array
          description: The list of returned creators in this collection.
          items:
            $ref: '#/components/schemas/CreatorSummary'
    CharacterList:
      properties:
        available:
          type: integer
          description: The number of total available characters in this list. Will always be greater than or equal to the "returned" value.
          format: int32
        returned:
          type: integer
          description: The number of characters returned in this collection (up to 20).
          format: int32
        collectionURI:
          type: string
          description: The path to the full list of characters in this collection.
        items:
          type: array
          description: The list of returned characters in this collection.
          items:
            $ref: '#/components/schemas/CharacterSummary'
    SeriesList:
      properties:
        available:
          type: integer
          description: The number of total available series in this list. Will always
            be greater than or equal to the "returned" value.
          format: int32
        returned:
          type: integer
          description: The number of series returned in this collection (up to 20).
          format: int32
        collectionURI:
          type: string
          description: The path to the full list of series in this collection.
        items:
          type: array
          description: The list of returned series in this collection.
          items:
            $ref: '#/components/schemas/SeriesSummary'
    StoryList:
      properties:
        available:
          type: integer
          description: The number of total available stories in this list. Will always
            be greater than or equal to the "returned" value.
          format: int32
        returned:
          type: integer
          description: The number of stories returned in this collection (up to 20).
          format: int32
        collectionURI:
          type: string
          description: The path to the full list of stories in this collection.
        items:
          type: array
          description: The list of returned stories in this collection.
          items:
            $ref: '#/components/schemas/StorySummary'
    CharacterSummary:
      properties:
        resourceURI:
          type: string
          description: The path to the individual character resource.
        name:
          type: string
          description: The full name of the character.
        role:
          type: string
          description: The role of the creator in the parent entity.
    EventSummary:
      properties:
        resourceURI:
          type: string
          description: The path to the individual event resource.
        name:
          type: string
          description: The name of the event.
    SeriesSummary:
      properties:
        resourceURI:
          type: string
          description: The path to the individual series resource.
        name:
          type: string
          description: The canonical name of the series.
    ComicSummary:
      properties:
        resourceURI:
          type: string
          description: The path to the individual comic resource.
        name:
          type: string
          description: The canonical name of the comic.
    Url:
      properties:
        type:
          type: string
          description: A text identifier for the URL.
        url:
          type: string
          description: A full URL (including scheme, domain, and path).
    CreatorSummary:
      properties:
        resourceURI:
          type: string
          description: The path to the individual creator resource.
        name:
          type: string
          description: The full name of the creator.
        role:
          type: string
          description: The role of the creator in the parent entity.
    StorySummary:
      properties:
        resourceURI:
          type: string
          description: The path to the individual story resource.
        name:
          type: string
          description: The canonical name of the story.
        type:
          type: string
          description: The type of the story (interior or cover).
    Image:
      properties:
        path:
          type: string
          description: The directory path of to the image.
        extension:
          type: string
          description: The file extension for the image.
    CharacterDataContainer:
      properties:
        offset:
          type: integer
          description: The requested offset (number of skipped results) of the call.
          format: int32
        limit:
          type: integer
          description: The requested result limit.
          format: int32
        total:
          type: integer
          description: The total number of resources available given the current filter
            set.
          format: int32
        count:
          type: integer
          description: The total number of results returned by this call.
          format: int32
        results:
          type: array
          description: The list of characters returned by the call.
          items:
            $ref: '#/components/schemas/Character'
    Character:
      properties:
        id:
          type: integer
          description: The unique ID of the character resource.
          format: int32
        name:
          type: string
          description: The name of the character.
        description:
          type: string
          description: A short bio or description of the character.
        modified:
          type: string
          description: The date the resource was most recently modified.
          format: date
        resourceURI:
          type: string
          description: The canonical URL identifier for this resource.
        urls:
          type: array
          description: A set of public web site URLs for the resource.
          items:
            $ref: '#/components/schemas/Url'
        thumbnail:
          $ref: '#/components/schemas/Image'
        comics:
          $ref: '#/components/schemas/ComicList'
        stories:
          $ref: '#/components/schemas/StoryList'
        events:
          $ref: '#/components/schemas/EventList'
        series:
          $ref: '#/components/schemas/SeriesList'
    CharacterDataWrapper:
      properties:
        code:
          type: integer
          description: The HTTP status code of the returned result.
          format: int32
        status:
          type: string
          description: A string description of the call status.
        copyright:
          type: string
          description: The copyright notice for the returned result.
        attributionText:
          type: string
          description: The attribution notice for this result.  Please display either
            this notice or the contents of the attributionHTML field on all screens
            which contain data from the Marvel Comics API.
        attributionHTML:
          type: string
          description: An HTML representation of the attribution notice for this result.  Please
            display either this notice or the contents of the attributionText field
            on all screens which contain data from the Marvel Comics API.
        data:
          $ref: '#/components/schemas/CharacterDataContainer'
        etag:
          type: string
          description: A digest value of the content returned by the call.
    ComicDataWrapper:
      properties:
        code:
          type: integer
          description: The HTTP status code of the returned result.
          format: int32
        status:
          type: string
          description: A string description of the call status.
        copyright:
          type: string
          description: The copyright notice for the returned result.
        attributionText:
          type: string
          description: The attribution notice for this result.  Please display either
            this notice or the contents of the attributionHTML field on all screens
            which contain data from the Marvel Comics API.
        attributionHTML:
          type: string
          description: An HTML representation of the attribution notice for this result.  Please
            display either this notice or the contents of the attributionText field
            on all screens which contain data from the Marvel Comics API.
        data:
          $ref: '#/components/schemas/ComicDataContainer'
        etag:
          type: string
          description: A digest value of the content returned by the call.
    ComicDataContainer:
        properties:
          offset:
            type: integer
            description: The requested offset (number of skipped results) of the call.
            format: int32
          limit:
            type: integer
            description: The requested result limit.
            format: int32
          total:
            type: integer
            description: The total number of resources available given the current filter
              set.
            format: int32
          count:
            type: integer
            description: The total number of results returned by this call.
            format: int32
          results:
            type: array
            description: The list of characters returned by the call.
            items:
              $ref: '#/components/schemas/Comic'
    Comic:
      properties:
        id:
          type: integer
          description: The unique ID of the character resource.
          format: int32
        digitalId:
          type: integer
          description: The ID of the digital comic representation of this comic. Will be 0 if the comic is not available digitally.
          format: int32
        title:
          type: string
          description: The canonical title of the comic.
        issueNumber:
          type: number
          format: double
          description: The number of the issue in the series (will generally be 0 for collection formats).
        variantDescription:
          type: string
          description: If the issue is a variant (e.g. an alternate cover, second printing, or director’s cut), a text description of the variant.
        description:
          type: string
          description: The preferred description of the comic.
        modified:
          type: string
          description: The date the resource was most recently modified.
          format: date
        isbn:
          type: string
          description: The ISBN for the comic (generally only populated for collection formats).
        upc:
          type: string
          description: The UPC barcode number for the comic (generally only populated for periodical formats).
        diamondCode:
          type: string
          description: The Diamond code for the comic.
        ean:
          type: string
          description: The EAN barcode for the comic.
        issn:
          type: string
          description: The ISSN barcode for the comic.
        format:
          type: string
          description: The publication format of the comic e.g. comic, hardcover, trade paperback.
        pageCount:
          type: integer
          description: The number of story pages in the comic.
          format: int32
        textObjects:
          type: array
          description: A set of descriptive text blurbs for the comic.
          items:
            $ref: '#/components/schemas/TextObject'
        resourceURI:
          type: string
          description: The canonical URL identifier for this resource.
        urls:
          type: array
          description: A set of public web site URLs for the resource.
          items:
            $ref: '#/components/schemas/Url'
        series:
          $ref: '#/components/schemas/SeriesSummary'
        variants:
          type: array
          description: A list of variant issues for this comic (includes the "original" issue if the current issue is a variant).
          items:
            $ref: '#/components/schemas/ComicSummary'
        collections:
          type: array
          description: A list of collections which include this comic (will generally be empty if the comic's format is a collection).
          items:
            $ref: '#/components/schemas/ComicSummary'
        collectedIssues:
          type: array
          description: A list of issues collected in this comic (will generally be empty for periodical formats such as "comic" or "magazine").
          items:
            $ref: '#/components/schemas/ComicSummary'       
        dates:
          type: array
          description: A list of key dates for this comic.
          items:
            $ref: '#/components/schemas/ComicDate'
        prices:
          type: array
          description: A list of prices for this comic.
          items:
            $ref: '#/components/schemas/ComicPrice'
        thumbnail:
          $ref: '#/components/schemas/Image'
        images:
          type: array
          description: A list of promotional images associated with this comic.
          items:
            $ref: '#/components/schemas/Image'
        creators:
          $ref: '#/components/schemas/CreatorList'
        characters: 
          $ref: '#/components/schemas/CharacterList'
        stories: 
          $ref: '#/components/schemas/StoryList'
        events:
          $ref: '#/components/schemas/EventList'